fix: going through missed days
