+++ src/common/constants.ts
@@ -4,3 +4,5 @@ export const ONE_DAY_IN_SECONDS = ONE_HOUR_IN_SECONDS * 24;
 export const ONE_WEEK_IN_SECONDS = ONE_DAY_IN_SECONDS * 7;
 
 export const MAX_FOLLOWERS_LIMIT = 5_000;
+
+export const SUCCESSFUL_CIO_SYNC_DATE = 'successful_cio_sync_date';
