+++ src/common/googleCloud.ts
@@ -3,6 +3,7 @@ import { PropsParameters } from '../types';
 import path from 'path';
 import { BigQuery } from '@google-cloud/bigquery';
 import { Query } from '@google-cloud/bigquery/build/src/bigquery';
+import { subDays } from 'date-fns';
 
 export const downloadFile = async ({
   url,
@@ -32,6 +33,7 @@ export const downloadJsonFile = async <T>({
   return JSON.parse(result);
 };
 
+// TODO: turn this into a get query function so we can pass the dates to supply the query
 export const userActiveStateQuery = `
   with d as (
     select uss.primary_user_id,
@@ -39,25 +41,25 @@ export const userActiveStateQuery = `
     min(registration_timestamp) as registration_timestamp,
 
       min(
-        case when period_end between date('2024-12-07' - interval 6*7 day) and '2024-12-07' then '1. active_last_6w'
-          when period_end between date('2024-12-07' - interval 12*7 day) and date('2024-12-07' - interval 6*7 + 1 day) then '2. active_7w_12w'
-          when date(u.last_app_timestamp) <  date('2024-12-07' - interval 12*7 day) then '3. active_12w+'
-          when date(u.registration_timestamp) <  date('2024-12-07' - interval 12*7 day) then '3. active_12w+'
+        case when period_end between date(@run_date - interval 6*7 day) and @run_date then '1. active_last_6w'
+          when period_end between date(@run_date - interval 12*7 day) and date(@run_date - interval 6*7 + 1 day) then '2. active_7w_12w'
+          when date(u.last_app_timestamp) <  date(@run_date - interval 12*7 day) then '3. active_12w+'
+          when date(u.registration_timestamp) <  date(@run_date - interval 12*7 day) then '3. active_12w+'
           else '4. never_active' end
       ) as previous_state,
 
       min(
-        case when period_end between date('2024-12-08' - interval 6*7 day) and '2024-12-08' then '1. active_last_6w'
-            when period_end between date('2024-12-08' - interval 12*7 day) and date('2024-12-08' - interval 6*7 + 1 day) then '2. active_7w_12w'
-            when date(u.last_app_timestamp) <  date('2024-12-08' - interval 12*7 day) then '3. active_12w+'
-            when date(u.registration_timestamp) <  date('2024-12-08' - interval 12*7 day) then '3. active_12w+'
+        case when period_end between date(@previous_date - interval 6*7 day) and @previous_date then '1. active_last_6w'
+            when period_end between date(@previous_date - interval 12*7 day) and date(@previous_date - interval 6*7 + 1 day) then '2. active_7w_12w'
+            when date(u.last_app_timestamp) <  date(@previous_date - interval 12*7 day) then '3. active_12w+'
+            when date(u.registration_timestamp) <  date(@previous_date - interval 12*7 day) then '3. active_12w+'
             else '4. never_active' end
       ) as current_state,
 
 
     from analytics.user as u
     left join analytics.user_state_sparse as uss on uss.primary_user_id = u.primary_user_id
-      and uss.period_end between date('2024-12-07' - interval 12* 7 day) and '2024-12-08'
+      and uss.period_end between date(@run_date - interval 12* 7 day) and @previous_date
       and uss.period = 'daily'
       and uss.app_active_state = 'active'
       and uss.registration_state = 'registered'
@@ -77,13 +79,20 @@ export const userActiveStateQuery = `
   where previous_state != current_state
 `;
 
-interface GetUsersActiveState {
+export const getUserActiveStateQuery = (runDate: Date): Query => {
+  const run_date = runDate.toISOString().split('T')[0];
+  const previous_date = subDays(runDate, 1).toISOString().split('T')[0];
+
+  return { query: userActiveStateQuery, params: { previous_date, run_date } };
+};
+
+export interface GetUsersActiveState {
   inactiveUsers: string[];
   downgradeUsers: string[];
   reactivateUsers: string[];
 }
 
-interface UserActiveState {
+export interface UserActiveState {
   current_state: string;
   previous_state: string;
   primary_user_id: string;
@@ -91,19 +100,18 @@ interface UserActiveState {
 
 const bigquery = new BigQuery();
 
-export const queryFromBq = async (
-  query: string,
-): Promise<UserActiveState[]> => {
-  const options: Query = { query };
-
-  const [job] = await bigquery.createQueryJob(options);
+export const queryFromBq = async (query: Query): Promise<UserActiveState[]> => {
+  const [job] = await bigquery.createQueryJob(query);
   const [rows] = await job.getQueryResults();
 
   return rows;
 };
 
-export const getUsersActiveState = async (): Promise<GetUsersActiveState> => {
-  const usersFromBq = await queryFromBq(userActiveStateQuery);
+export const getUsersActiveState = async (
+  runDate: Date,
+): Promise<GetUsersActiveState> => {
+  const query = getUserActiveStateQuery(runDate);
+  const usersFromBq = await queryFromBq(query);
   const inactiveUsers: string[] = [];
   const downgradeUsers: string[] = [];
   const reactivateUsers: string[] = [];
